requirements.txt
streamlit>=1.28.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
plotly>=5.17.0
python-dateutil>=2.8.2
bcrypt>=4.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
